Add ferris sweep layout v4
