+++ keyboards/ferris/keymaps/hbmorrisonv3/config.h
@@ -39,9 +39,6 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 // Enable key repeat
 #define QUICK_TAP_TERM 100
 
-// Shorten leader timeout
-#define LEADER_TIMEOUT 200
-
 // Underglow configuration
 #ifdef RGBLIGHT_ENABLE
 #    define RGBLIGHT_EFFECT_BREATHING
