+++ keyboards/ferris/keymaps/hbmorrisonv3/rules.mk
@@ -1,4 +1,3 @@
 TAP_DANCE_ENABLE = yes
 MOUSEKEY_ENABLE = yes
 SEND_STRING_ENABLE = yes
-LEADER_ENABLE = yes
